requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# mongo/data.py: 1
Faker == 8.1.0

# mongo/consumer.py: 3
# mongo/producer.py: 1
kafka_python == 2.0.2

# mongo/consumer.py: 4
pymongo == 3.11.3
